Backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
botocore==1.37.7
certifi==2025.1.31
cffi==1.17.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
inflection==0.5.1
jmespath==1.0.1
Markdown==3.7
minio==7.2.15
packaging==24.2
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
mysqlclient==2.1.1